Added tests for upcoming AnyFormat + some refactoring
