+++ src/test/scala/com/trueaccord/scalapb/json/StructFormatSpec.scala
@@ -5,10 +5,7 @@ import com.google.protobuf.struct._
 import com.trueaccord.scalapb._
 import jsontest.test3.StructTest
 
-class StructFormatSpec extends FlatSpec with MustMatchers {
-  val JavaJsonPrinter = com.google.protobuf.util.JsonFormat.printer()
-  val JavaJsonParser = com.google.protobuf.util.JsonFormat.parser()
-
+class StructFormatSpec extends FlatSpec with MustMatchers with JavaAssertions {
   val ListValueExample = ListValue(values = Seq(
     Value(Value.Kind.NumberValue(-245.0)),
     Value(Value.Kind.BoolValue(true)),
@@ -26,22 +23,6 @@ class StructFormatSpec extends FlatSpec with MustMatchers {
     "f2" -> Value(Value.Kind.StructValue(StructExample)),
     "f3" -> Value(Value.Kind.NullValue(NullValue.NULL_VALUE))
     ))
-
-  def assertJsonIsSameAsJava[T <: GeneratedMessage with Message[T]](v: T)(
-    implicit cmp: GeneratedMessageCompanion[T]) = {
-    val scalaJson = com.trueaccord.scalapb.json.JsonFormat.toJsonString(v)
-    val javaJson = JavaJsonPrinter.print(
-        cmp.asInstanceOf[JavaProtoSupport[T, com.google.protobuf.GeneratedMessageV3]].toJavaProto(v))
-
-    import org.json4s.jackson.JsonMethods._
-    parse(scalaJson) must be (parse(javaJson))
-    JsonFormat.parser.fromJsonString[T](scalaJson) must be(v)
-  }
-
-  def javaParse[T <: com.google.protobuf.GeneratedMessageV3.Builder[T]](json: String, b: com.google.protobuf.GeneratedMessageV3.Builder[T]) = {
-    JavaJsonParser.merge(json, b)
-    b.build()
-  }
 
   "Empty value" should "be serialized to null" in {
     JavaJsonPrinter.print(com.google.protobuf.Value.newBuilder().build()) must be ("null")
